split integration test into different files
